Windows compatibility issues.

+++ trunk/src/flitr/modules/flitr_image_processors/average_image/fip_average_image.cpp
@@ -116,7 +116,7 @@ bool FIPAverageImage::trigger()
             const size_t componentsPerLine=componentsPerPixel * width;
             
             //Update this slot's average image here...
-            size_t y=0;
+            int y=0;
             {
 #ifdef _OPENMP
                 //omp_set_num_threads(4);//There might be some blocking operation or mem bandwidth limited code??? because the parallel for seems to work best with much more threads than CPU cores.
@@ -129,7 +129,7 @@ bool FIPAverageImage::trigger()
 #ifdef _OPENMP
 #pragma omp for nowait
 #endif
-                    for (y=0; y<height; y++)
+                    for (y=0; y<(int)height; y++)
                     {
                         const size_t lineOffset=y * componentsPerLine;
