+++ trunk/src/flitr/modules/flitr_image_processors/gradient_image/fip_gradient_image.cpp
@@ -93,7 +93,7 @@ bool FIPGradientXImage::trigger()
             const size_t width=imFormat.getWidth();
             const size_t height=imFormat.getHeight();
             
-            size_t y=0;
+            int y=0;
             {
 #ifdef _OPENMP
                 //omp_set_num_threads(4);//There might be some blocking operation or mem bandwidth limited code??? because the parallel for seems to work best with much more threads than CPU cores.
@@ -106,7 +106,7 @@ bool FIPGradientXImage::trigger()
 #ifdef _OPENMP
 #pragma omp for nowait
 #endif
-                    for (y=0; y<height; y++)
+                    for (y=0; y<(int)height; y++)
                     {
                         const size_t lineOffset=y * width;
                         
@@ -200,7 +200,7 @@ bool FIPGradientYImage::trigger()
             const size_t width=imFormat.getWidth();
             const size_t height=imFormat.getHeight();
             
-            size_t y=0;
+            int y=0;
             {
 #ifdef _OPENMP
                 //omp_set_num_threads(4);//There might be some blocking operation or mem bandwidth limited code??? because the parallel for seems to work best with much more threads than CPU cores.
@@ -213,7 +213,7 @@ bool FIPGradientYImage::trigger()
 #ifdef _OPENMP
 #pragma omp for nowait
 #endif
-                    for (y=1; y<height; y++)
+                    for (y=1; y<(int)height; y++)
                     {
                         const size_t lineOffset=y * width;
